Woops, broken again...
But this is something that beginning parser writers, or even me right
now trying to find a clean line would try. The parsing chain *should*
find that, <expression> expands to <assignment-expression> expands
to <comparison-expression> expands to match our statement, so...
why doesn't it work?

Well, it could just be that it's in the wrong place. Let's move it up
one step in the parsing chain and see what happens...