Fixed it, whew.
It still looks a bit weird, though, for several reasons. First off,
it's got two operators, '**' and '<='. If we keep "fixing" things this
way we'll end up with an exponential explosion over exponentiations.
Not a good look.

So, let's try a different approach. Right now we have this:

  rule assignment-expression
    {
    | <variable-name> '**' \d+ '<=' <expression>
    | <variable-name> '=' <expression>
    | <comparison-expression>
    }

  rule comparison-expression
    {
    <variable-name> '<=' <expression>
    }

  rule expression
    {
    | <assignment-expression>
    # ...
    }

It looks like the assignment doesn't really belong in that list at all,
so let's refactor it out, plus it'd make sense to say:

  <variable> '=' <some-expression>...

  rule assignment-expression
    {
    | <variable-name> '**' \d+ '<=' <expression>
    | <comparison-expression>
    }

  rule comparison-expression
    {
    <variable-name> '<=' <expression>
    }

  rule expression
    {
    | <assignment-expression>
    | <variable-name> '=' <expression>
    # ...
    }

But now there are no assignments in <assignment-expression>, let's fix
that next.