Cleaning up a tiny issue first...
We want to be able to parse 'I <= 10' and '2 <= I', obviously. The
latter form may look a little weird, but it's perfectly legal, and
you just *know* that if we forget that case, then someone's going
to come along, use that and complain that the grammar's broken.

So let's fix that. Instead of requiring the LHS (left-hand side) of our
'<=' operator to be a variable name, let's change it to accept either
a variable name or a number.

(I'm deliberately ignoring negative numbers for the moment.)

And for this I'll borrow a little bit of compiler terminology that may
make more sense now that you've seen it in a few (dozen) Perl 6 syntax
error messages, 'term'.

This is a deliberately generic word, and it's taken from decades of
compiler research. Basically the first person that wrote a paper on
grammars realized "Hrm, the left-hand side of '<=' can take either a
number or a variable. I can't call the term a <number> because that's
not all, I can't call it a <variable>, because that's not.. wait, term.
That works out."

So we use 'term' for this name, mostly for historic reasons, and as a
side benefit, when you go to read Wikipedia about this section, that
is a term (ha!) that they'll use there too.

You'll see it usually in <E> + <T>, which is <expression> '+' <term> and
we'll get there soon enough. But I'm doing stepwise reduction, so this
all comes in nice bite-sized pieces.