Sneaking up on it...
Let's replace the LHS of the '<=' operator with <term>. Now, this
should look very familiar from the last commit message, where I was
talking about terms and such. Compiler writers and optimizers like
simple things like:

<term> + <expression>

because they're fast to parse. But I like it because you can now see
the commonality here:

  rule math-expression
    {
    | <variable-name> '**' <term> '<=' <expression>
    | <comparison-expression>
    }

  rule comparison-expression
    {
    <term> '<=' <expression>
    }

You'll see "<term> '<=' <expression>" in two places. That means we
can replace one with the other, like so:

  rule math-expression
    {
    | <variable-name> '**' <comparison-expression>
    | <comparison-expression>
    }

  rule comparison-expression
    {
    <term> '<=' <expression>
    }

And we're almost there.