And finally, we can see what's going on...
  rule math-expression
    {
    | <term> '**' <comparison-expression>
    | <comparison-expression>
    }

  rule comparison-expression
    {
    <term> '<=' <expression>
    }

  rule expression
    {
    | <math-expression>
    | <variable-name> '=' <expression>
    # ...
    }

So from the bottom up (expression), this can be read as:

An expression can either be math or assignment.
If it's math, it can either be X to the exponent of a ... comparison
expression? That... seems wrong to me. Now we have a subtler issue on
our hands.

'I**2 <= N' parses to:

<term = 'I'> '**' ( <term = '2'> '<=' ( <variable-name ) )

This is harder to see, but the parser looks for the math-expression first,
and then the comparison expression. So instead of parsing to

( ( I ** 2 ) <= N )

our statement parses to:

( I ** ( 2 <= N ) )

which is most definitely not what we want. Unfortunately it's hard to see
why this is the case without delving into the parse tree. I'll see about
adding a debugging utility to show this problem on the next commit.