Voila', a quick utility to demonstrate the parser issue...
Check out show-expression-tree.pl6 in the main directory. This may
get moved into the utilities later on once it gets fancified, but
for now, here's how to read the output:

05:48 $ perl6 show-expression-tree.pl6
｢I**2 <= N｣
 math-expression => ｢I**2 <= N｣
  term => ｢I｣
   variable-name => ｢I｣
  comparison-expression => ｢2 <= N｣
   term => ｢2 ｣
    variable-name => ｢2｣
   expression => ｢N｣
    function-call => ｢N｣
     function-name => ｢N｣

"I**2 <= N" is a math expression.
Its LHS ... In fact, let's do one little change to make this look
nicer, and introduce aliases.