Whew. Even more verbose, but...
Here's the sample output: (again, edited for brevity)

05:52 $ perl6 show-expression-tree.pl6
｢I**2 <= N｣
 math-expression => ｢I**2 <= N｣
  LHS => ｢I｣
  0 => ｢**｣
  comparison-expression => ｢2 <= N｣
   LHS => ｢2 ｣
   0 => ｢<= ｣
   RHS => ｢N｣
  RHS => ｢2 <= N｣
   LHS => ｢2 ｣
   0 => ｢<= ｣
   RHS => ｢N｣

So, let's rewrite this into how Perl 6 will look at things.

math-expression => {
  LHS      => 'I',
  operator => '**',
  RHS      => {
    LHS      => '2 ',
    operator => '<= ',
    RHS      => 'N'
  }
}

So, it would indeed parse out to ( I ** ( 2 <= N ) )

Which is definitley not the reading order we'd prefer. This is a
precedence bug. '**' should have very "tight" precedence, which means
that we want it at the bottom of our tree, so we want the tree to look
like this:

comparison-expression => {
  LHS => {
    LHS      => 'I',
    operator => '**',
    RHS      => '2 '
  },
  operator => '<= ',
  RHS      => 'N'
}

It looks like we have to swap something... and it could be a problem.
For that layout to work, we'd have to make the first term of
<comparison-expression> be an <expression>, and we can't let that
recurse like we were doing before. This has already infinilooped in its
history often enough, and I'd like to avoid that just in case people
aren't prepared.