Collect the appropriate stuff into <term>
I know it's a bit of a change from the last commit, but I decided
to clean up the term to what it should be. You'll also note that in
doing so I do introduce a bit of recursion, especially via the
<array> term.

Arrays are now just lists of expressions,
so term -> array -> expression -> term is now a potential loop we
have to watch out for. It adds a bit more challenge, but I think it
makes the expressions just a bit cleaner.