Some grammar cleanup
You guessed it, I'm procrastinating on figuring out the best way to
transition what I've got to a proper expression grammar. But cleaning
little things up still has a value, because we now have some more use
cases to play with.

I've tweaked <variable-name> to add the '.length' modifier on it,
tweaked the definition of <array> to allow '[]' as a legitimate empty
array, and added an optional <comment>* before the main program body.

I'm afraid that I may have to rewrite the entire thing, but there is
a simple way to get started. I may add a temporary test tool to show
what's going on in more detail, but we'll see.