Playing in the sandbox
I've added test-expression.pl6 as a sandbox for us to work in, temporarily.
The main file is really cluttered, and it feels good to be able to
focus on just the expression. I probably should have started with this
approach, but I thought that watching the existing code vanish would be
satisfying.

For the time being, run test-expression.pl6 to check the progress of what
we're working on. You'll notice that a *lot* of the tests right now fail.
We're going to rebuild the expression, and then add it back in to the
main grammar. A different approach, but it'll work out.

At the moment, you'll notice that only the variables and constants
pass the test. Function calls aren't even in here yet, and those are
simple. In fact, starting out with that might be a good place.

I'm not going to remove this file, by the way. It'd confuse people reading
the log later, and it's also a good way to show that hey, I too get stuck.