Add the full test-expression file
As I said before, I'm not going to remove these because it helps to
show my thought processes, and especially in this case it's a good way
to show what's at the core of the language without any of the clutter
that goes with the main file.

Let's play a bit, and try to reformulate the expression without any of
the clutter. I've taken care of collecting all of the expressions that I
can see, and put them in the test file.

One thing to note is that I made sure to add one of each of the basic
classes of operation, plus one that doesn't yet exist in my sample
files. I've got '==' and '<=' for comparison, '^' for bitwise, '*' and
'+' for math, and '**' for exponents. Also '('..')' just in case.

I've done this mainly to be sure that I cover the common "precedence"
classes. You might be familiar with the PEMDAS mnemonic of
'parentheses', 'exponents', 'mult', 'div', 'add', 'subtract' order.
This way I can be sure that the parsed output nests correctly. Also,
you'll see that I've got a single function that does all the parsing,
this is handy because it's a single place to add a quick 'say $g.parse(...)'
so you can see more detail.