Fix the parentheses...
And add a test to show that it indeed works. We're going to follow
the PEMDAS acronym on the way down, and put logic operators below AS
in the acronym stack.

This just means that:

<expression> => P(arentheses)
P(arentheses) => E(xponentiation)
E(xponentiation) => M(ultiplication), D(ivision)
M(ultiplication), D(ivision) => A(ddition), S(ubtraction)
A(ddition), S(ubtraction) => L(ogic)

We're going to try to follow this pattern on the way down.