Add exponentiation, move parentheses around
Exponentiation works, and has the proper left-to-right precedence, in
that '1**2**3' parses the same as '1**(2**3)'. It's an easy check for
you to make, just add 'say ' to the start of the '$g.parse(...)' line
in your test script.

So, parentheses and exponentiation finally get to rumble, and it all
goes down passing the tests with flying colors. Yay. The important
tests are (2**3)**4 and 2**(3**4). These have to be different, because
they're definitely mathenatically different. (2**3)**4 == 8**4, and
2**(3**4) == 2**81, and 4096 <definitely>!= 2417851639229258349412352.

So they need to parse differently, and they do. Yay, it's time to move
on and handle multiplication/division next.