Woo, adding in multiplication/division was simple.
Just a reminder, we're still in test-expression.pl6, not the main test
file. Getting closer, though. You should be able to see now how to
add more layers, but I'm not going to speed ahead. We're pretty close
now, though.

The testing is a bit skimpy, but I'm relatively confident about the
progress. We're adding the '*' operator to the original '**' operator,
so I want to make sure to cover all the combinations of '*', especially
'***' and '1***1' because while I don't think they'll break, the
purpose of testing is to make *sure* they don't.

We could put this back in the grammar, but I don't think we'd gain
anything yet.