Woops, broke something... again.
But my test script catches it. As a reminder, we're still on
the test-expression.pl6 file. Run that and you'll see that most of the
tests run, but '1-2*3' breaks along with one other one.

Addition/Subtraction should look almost exactly like mult/division.
But it doesn't quite do that. If you trace the grammar by hand, you
can see that...

   multiply-expression
   term
   |
   V
  '1-2*3'

... and we're stuck. We should be able to transition to add-expression,
but we're not. The simple answer is that I've "accidentally" left the
<term> intact. It should descend to <add-expression>, not <term> because
that's how the pattern we've set up works. We'll fix that and go on.

I just left this breakage in because I did it by accident, and stopped
myself before "fixing" it so you can see the progress.