Getting simpler as we go along...
This took maybe a few minutes of editing. You'll see that we finally
can parse 'I**2 <= N' which is in our real test file as well. At last,
a "real world" application.

All we need to handle is function calls, and that we can probably do outside
the grammar. So, let's merge this back into the main grammar.