Add ^ (probably xor) as an operator.
It just occurred to me that we don't have unary operators... yet.
In any case, now that we have things sort of structured, it should
be obvious how to add an operator in. The first thing to figure out
is where it fits in the precedence hierarchy, and I'm going to assume
that xor binds very tightly.

Also, I'm going to guess that the operator names here follow C
conventions, and add '&' and '|' to the list. Next thing is to add
binding operators, then our first really big expressions.