Move <binding> out of the expressions.
This is mainly because I don't believe that the LHS of a "binding"
(just making up that name because Perl 6 has something similar)
should be an expression, and that could get in the way of things.

I've noticed a similar problem with <thingie> that will be harder
to fix, but again, I'd like to remind people that this is purely
an experiment in stream-of-consciousness programming. It was never
my intent to write things so that each step was pristine and
the (usually bad) decisions I took were all magically correct.