Starting to change my mind on 'thingie's.
Specifically there's a problem in that term that I'd like to
nip in the bud. And this proves beyond a shadow of a doubt that I'm
sort of making this stuff up on the fly. It actually allows
<expression> <expression> because <comma> is an option, so it's
probably a *bad* thing.

I'm going to take a little while to properly do the expressions,
so we'll shift gears here a bit. The first step is to rewrite
the function-definitions to properly use ',' as a separator instead
of just a generic term.

This also speaks to the Reddit thread that started about this problem.
I'm still not a great fan of using whitespace-sensitive constructs
because it's too easy for a new user to come along and see what looks
like an "untidy construct", "fix" it and later on when there's a
regression they search wildly for the answer and don't find it
because the problem they accidentally caused was to *their* eyes a
little esthetic issue, already forgotten.

I'm not at all putting down the Reddit thread, it just happens that
while I agree with the spirit of the fix, I don't happen to agree
with the letter.