And I even have a sample of why it broke...
Check out this function definition down at the bottom...

<function-name> '=>'
   'foreach(Len in 1..15)'
      <thingie> <comma> <comment>*
      <thingie> <comma> <comment>*
      <thingie> <comma> <comment>*
      <thingie> <comma> <comment>*
      'time(' <variable-name> '=' <expression>  ')' <comma> <comment>*
      <thingie> <comment>*
   'end'
<period>

I've had that commented-out line in there probably since I created
the <thingie> construct, and it was never actually used. It was just
parsing as an empty string, yet I've been carrying it along through
the entire journey.

This is *precisely* why I wanted to fix this sort of problem, I'm
honestly pleased that I even had an example of why this was a
problem, I can assure you that it was purely happenstance.

Now, let's give the <thingie> a proper name, and we can go back to
the big blocks.