Whew, finally we can call 'thingie' a <statement>
Which is what it was supposed to be all along, really.

Things like foreach loops now read a lot better. A foreach loop
has a declaration, a list of statements separated by commas, and
then an 'end' term. Easy peasy.

Also <statement> looks visually simpler, although I'll bet we could
get rid of that last <comment> block if we tried. But I'm not going
to worry about that for now, we're close to getting rid of almost all
the file.