And all the <function-declarations> are collapsed!
All that's left to do is clean up the <comment> blocks, collapse
the <function-definition>s and <fact-definition>s into the main
file and we've probably parsed a good portion of picat.

Although I've looked at the grammar and there's still lots to do,
most of it slots in to the rules we've put together.