Playing with the expression hierarchy
I'm not going to go through each change in detail, but what I've done
is change the expression hierarchy so it's closer to what the actual
Picat grammar is. And add some stages so we get more of what a full
Picat expression can capture.

Here's a quick example of how to add in a class of operators. Let's
assume things looked like this, as they did at one point:

  rule assignment-expression
    {
    | <variable-name> '=' <expression>
    | <exponent-expression>
    }

  rule exponent-expression
    {
    <shift-expression> [ '**' <shift-expression> ]*
    }

This means that 'I = X ** Y + 1' would get parsed as:

'(I = ( X ** ( Y + 1 ) ) )'

where '**' is the "loosest" precedence. We've just been through the
grammar and see "oh crud, there's a "concat" operator tighter than
exponentiation, so we need to parse that as the very next operator.

First we create the new rule like so:

  rule assignment-expression
    {
    | <variable-name> '=' <expression>
    | <exponent-expression>                               >----+
    }                                                          |
                                                               |
  # Here's the new rule, notice how we make its "children" the |
  # <exponent-expression> below.                               |
  #                                                            |
  rule concat-expression                                       |
    {                                                          |
    <exponent-expression> [ '++' <exponent-expression> ]* >--+ |
    }                                                        | |
                                                             | |
  rule exponent-expression                                <--+-+
    {
    <shift-expression> [ '**' <shift-expression> ]*
    }

Then make the original "top" rule "descend" to the new rule instead
of the exponent rule;

  rule assignment-expression
    {
    | <variable-name> '=' <expression>
    | <concat-expression>                                 >-+
    }                                                       |
                                                            |
  rule concat-expression                                  <-+
    {
    <exponent-expression> [ '++' <exponent-expression> ]* >--+
    }                                                        |
                                                             |
  rule exponent-expression                                <--+
    {
    <shift-expression> [ '**' <shift-expression> ]*
    }

Like so. Just a useful editing tip.