Add the definition of <predicate-directive> from the original
The grammar says that things like 'index' and 'table' are bound
to what we're temporarily calling a function, thoguh we should
rename that later on.

Most of this comes from the grammar, but watch in the next commit
what happens when we use this term through the rest of the
code block.